Rename folder to match path
