+++ front/app/api/home_page/types.ts
