+++ front/app/api/home_page/useHomepageSettings.ts
