+++ front/app/api/home_page/useUpdateHomepageSettings.ts
