+++ front/app/containers/Admin/pagesAndMenu/EditHomepage/BottomInfoSection/index.tsx
@@ -1,11 +1,11 @@
-import useHomepageSettings from 'api/homepage_settings/useHomepageSettings';
+import useHomepageSettings from 'api/home_page/useHomepageSettings';
 import React from 'react';
-import { IHomepageSettingsAttributes } from 'api/homepage_settings/types';
+import { IHomepageSettingsAttributes } from 'api/home_page/types';
 import { useIntl } from 'utils/cl-intl';
 import { isNilOrError } from 'utils/helperUtils';
 import { homeBreadcrumb } from '../../breadcrumbs';
 import GenericBottomInfoSection from 'containers/Admin/pagesAndMenu/containers/GenericBottomInfoSection';
-import useUpdateHomepageSettings from 'api/homepage_settings/useUpdateHomepageSettings';
+import useUpdateHomepageSettings from 'api/home_page/useUpdateHomepageSettings';
 
 const BottomInfoSection = () => {
   const { formatMessage } = useIntl();
