+++ front/app/containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/CTASettings.test.tsx
@@ -2,7 +2,7 @@
 // @ts-nocheck
 import React from 'react';
 import { render, screen } from 'utils/testUtils/rtl';
-import { IHomepageSettingsAttributes } from 'api/homepage_settings/types';
+import { IHomepageSettingsAttributes } from 'api/home_page/types';
 
 import CTASettings from './CTASettings';
