+++ front/app/containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/CTASettings.tsx
@@ -10,7 +10,7 @@ import {
   CTASignedInType,
   CTASignedOutType,
   IHomepageSettingsAttributes,
-} from 'api/homepage_settings/types';
+} from 'api/home_page/types';
 import styled from 'styled-components';
 import { CLErrors, Multiloc } from 'typings';
 import { FormattedMessage } from 'utils/cl-intl';
