+++ front/app/containers/Admin/pagesAndMenu/containers/GenericHeroBannerForm/HeaderImageDropzone.tsx
@@ -15,7 +15,7 @@ import {
 import {
   IHomepageSettingsAttributes,
   THomepageBannerLayout,
-} from 'api/homepage_settings/types';
+} from 'api/home_page/types';
 import { UploadFile } from 'typings';
 
 const HeaderImageOverlay = styled.div<{
