+++ front/app/containers/HomePage/SignedOutHeader/FixedRatioLayout.tsx
@@ -11,7 +11,7 @@ import {
 } from 'components/LandingPages/citizen/FixedRatioLayout';
 import HeaderContent from './HeaderContent';
 
-import { IHomepageSettingsData } from 'api/homepage_settings/types';
+import { IHomepageSettingsData } from 'api/home_page/types';
 
 interface Props {
   homepageSettings: IHomepageSettingsData;
